requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
multidict==6.0.4
python-telegram-bot==20.7
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
urllib3==2.1.0
yarl==1.9.4
